requirements.txt
openai
langchain
faiss-cpu
sentence-transformers
streamlit
matplotlib
pandas
